Separated code in on_join on the server into clear join lobby and join other room sections. It can be pared down even more.
Currently lobby remembers username because it's
stored in flask session - this may change, it
might make sense to store it under a temp_username
key in the flask session instead of just username.

OR to not store it at all in the flask session as
the username is supposed to be able to change from
room to room.

Anyway it appears to be stable for now, will focus
more on visuals and styling