Updated app to be able to display rooms depending on choice on index.html.
Unintended consequence / BUG: when creating a room
for a game of lobby user is not currently in,
lobby type will change to the type of the newly
created room and transport the user to that lobby
instead.

Planning to change create room to only create
rooms of chosen game type which would probably
prevent this from happening, but still not sure
why it's happening.