I think cause of rejoin bug was player sid was not being updated in the game state. Removed all sids from game state so would not have to update them; using sids from room.users.
This is slightly messy because the users are not
in a dict, so every time I need to find a user for
their name or their sid I use a loop. May need to
make into a dict so sid can be accessed by name.

Fix needs to be tested.