Ready to merge card_animation branch with main branch!
Fixed end game bug in thirty_one_game.py

Set game grids to have fixed sizes so they don't
keep resizing when a card is added / removed

Re-centered board grid that had become un-centered

Eventually I want to populate the grid to more
closely resemble a real game table where self is
at the bottom and others are around equidistant
around the table.

The issue with this is players are being added to
the grid as they join, which means they would have
to be rearranged when game starts.

Number of players must be known to achieve this
smoothly. Could add players to a list on the
screen and then players are distributed around
'table' when game is started.