Changed rank of 10 to 'T' to standardize across project and not cause any translation issues.
Removed unzip and zip card because unzipping would
require passing in rank_to_value dict every time,
and it was only used in 2 places. I replaced those
uses with a lookup to the rank_to_value dict
instead of creating a whole card object, as well
as using the card str (portable) format when
possible.